move geolocation to objects
